compat: fix pointer→Int truncation and double narrowing on 64-bit/Clang
W3DProgressBar.cpp:
- winGetUserData() returns void*; cast through uintptr_t before truncating
  to Int to silence Clang 64-bit pointer-size error (same pattern as
  W3DModelDraw.cpp fix)

CMakeLists.txt:
- Add -Wno-narrowing for gameenginedevice on non-MSVC compilers.
  W3DMainMenu.cpp / W3DControlBar.cpp multiply Int screen coordinates by
  double literals and store them back into Int struct fields — legitimate
  screen-coordinate arithmetic that Clang C++11 flags as narrowing.

+++ GeneralsMD/Code/GameEngineDevice/Source/W3DDevice/GameClient/GUI/Gadget/W3DProgressBar.cpp
@@ -76,7 +76,7 @@ void W3DGadgetProgressBarDraw( GameWindow *window, WinInstanceData *instData )
 {
 	ICoord2D origin, size, start, end;
 	Color backColor, backBorder, barColor, barBorder;
-	Int progress = (Int)window->winGetUserData();
+	Int progress = (Int)(uintptr_t)window->winGetUserData();
 
 	// get window size and position
   window->winGetScreenPosition( &origin.x, &origin.y );
@@ -186,7 +186,7 @@ void W3DGadgetProgressBarImageDrawA( GameWindow *window, WinInstanceData *instDa
 {
 	ICoord2D origin, size;
 	const Image *barCenter, *barRight, *left, *right, *center;
-	Int progress = (Int)window->winGetUserData();
+	Int progress = (Int)(uintptr_t)window->winGetUserData();
 	Int xOffset, yOffset;
 	Int i;
 	// get window size and position
@@ -229,7 +229,7 @@ void W3DGadgetProgressBarImageDraw( GameWindow *window, WinInstanceData *instDat
 	ICoord2D origin, size, start, end;
 	const Image *backLeft, *backRight, *backCenter, 
 				 *barRight, *barCenter;//*backSmallCenter,*barLeft,, *barSmallCenter;
-	Int progress = (Int)window->winGetUserData();
+	Int progress = (Int)(uintptr_t)window->winGetUserData();
 	Int xOffset, yOffset;
 	Int i;
